requirements.txt
# ==============================================
# GROCERY STORE + WHATSAPP BOT - DEPENDENCIES
# ==============================================

# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Async & Task Queue
celery>=5.3.4
redis>=5.0.1
django-redis>=5.4.0

# Server
gunicorn>=21.2.0
uvicorn>=0.24.0
whitenoise>=6.6.0

# WhatsApp API
httpx>=0.25.2
aiohttp>=3.9.1

# Speech-to-Text
google-cloud-speech>=2.21.0
openai>=1.3.7
openai-whisper
setuptools-rust
torch

# Cloud Storage
boto3>=1.33.6
django-storages>=1.14.2
cloudinary>=1.36.0

# Image Processing
Pillow>=10.1.0
python-magic>=0.4.27

# PDF Generation
reportlab>=4.0.7
xhtml2pdf>=0.2.11

# Security & Validation
python-decouple>=3.8
cryptography>=41.0.7

# Monitoring & Logging
sentry-sdk>=1.38.0

# Utilities
phonenumbers>=8.13.26
python-dateutil>=2.8.2
requests>=2.31.0

# Development (remove in production)
django-debug-toolbar>=4.2.0
ipython>=8.18.1

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
